fix: async error centralization
